ビルドエラー修正

+++ client/src/main.rs
@@ -97,8 +97,8 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         let width = args.width as u32;
         let height = args.height as u32;
         let renderer = sjrt::PathTracerEx::default()
-            .with_depth(args.depth_max)
-            .with_sampling_count(args.sampling_count);
+            .with_depth(args.depth_max as u32)
+            .with_sampling_count(args.sampling_count as u32);
         let rays = sjrt::Camera::builder()
             .with_field_of_view(std::f32::consts::PI / 5.5)
             .with_position(&nalgebra::Vector3::new(4.8, 4.73, -8.0))
